Extract transport implementations into separate files
